% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eztune.R
\docType{package}
\name{EZtune-package}
\alias{EZtune}
\alias{EZtune-package}
\title{Tuning supervised learning models.}
\description{
The EZtune package contains two functions and four datasets. The functions
are eztune and eztune.cv. The datasets are lichen, lichenTest, mullein, and
mulleinTest
}
\section{Eztune functions}{

The EZtune functions are eztune and eztune.cv. The function eztune will
autotune an adaboost, gradient boosting machine, and support vector
machine using a genetic algorithm or a quasi-Newton-Raphson optimizer.
The eztune.cv function returns the cross-validated error for an object
produced by eztune.
}

\section{EZtune datasets}{

The lichen dataset has 840 observations and 40 variables. Seven of the
variables are binary response variables. The lichenTest dataset is a test
dataset for the lichen dataset that consists of 300 observations and
the same 40 variables.

The mullein dataset consists of 12094 observations and 32 variables. One
variable is a binary response. The mulleinTest dataset is a test
dataset for the mullein data that consists of 1512 observations
and the same 32 variables as mullein.
}

\author{
\strong{Maintainer}: Jill Lundell \email{jflundell@gmail.com}

}
